+++ tests/frontend/org/voltdb/benchmark/tpcc/procedures/neworder.java
@@ -144,7 +144,7 @@ public class neworder extends VoltProcedure {
         return -1;
     }
 
-    public VoltTable[] run(short w_id, byte d_id, int c_id, TimestampType timestamp, int[] item_id, byte[] supware, int[] quantity) throws VoltAbortException {
+    public VoltTable[] run(short w_id, byte d_id, int c_id, TimestampType timestamp, int[] item_id, short[] supware, int[] quantity) throws VoltAbortException {
         assert item_id.length > 0;
         assert item_id.length == supware.length;
         assert item_id.length == quantity.length;
